Fixed scrolledwindow padding
In some cases (e.g. Gnome Software Center > intalled apps) the widget
contained inside a scrolledwindow is so close to the top of the container
that overlaps the scrolledwindow box-shadow.

Giving 1px top padding to save some space for scrolledwindow box-shadow,
without altering too much the behavior of the widget is other use cases.

see #465